# mvcc: chunk reads for restoring