Loading all keys at once would cause etcd to use twice as much
memory than it would need to serve the keys, causing RSS to spike on
boot. Instead, load the keys into the mvcc by chunk. Uses pipelining
for some concurrency.

Fixes #7822

tmpfs benchmarks:

Before:
```
go test -v -run=BenchmarkStoreRestore -bench=BenchmarkStoreRestore
BenchmarkStoreRestoreRevs1-8     1000000      4446 ns/op     814 B/op      7 allocs/op
BenchmarkStoreRestoreRevs10-8     200000     11053 ns/op    5397 B/op     38 allocs/op
BenchmarkStoreRestoreRevs20-8     100000     17458 ns/op   10518 B/op     69 allocs/op
```

After:
```
go test -bench=BenchmarkStoreRestore -run=BenchmarkStoreRestore
BenchmarkStoreRestoreRevs1-8    	 1000000	      2464 ns/op	     641 B/op	       6 allocs/op
BenchmarkStoreRestoreRevs10-8   	  200000	      7396 ns/op	    4924 B/op	      37 allocs/op
BenchmarkStoreRestoreRevs20-8   	  100000	     13664 ns/op	    9715 B/op	      68 allocs/op
```