@heyitsanthony 2017-05-25
@davissp14 can you confirm this fixes the RSS spike?