@davissp14 2017-06-01
Just confirming this did fix the RSS spike.  Thanks guys