@heyitsanthony 2017-06-01
OK, I'm fine with moving this up to 3.2.

@gyuho @xiang90 any objections to slipping this into 3.2?